requirements.txt
Flask>=3.0.0
Flask-Limiter>=3.5.0
Flask-JWT-Extended>=4.5.3
celery>=5.3.6
redis>=5.0.1
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.23
pydantic>=2.5.2
langchain>=0.1.0
langchain-core>=0.1.0
tavily-python>=0.3.0
sentence-transformers>=2.2.2
transformers>=4.36.2
torch>=2.1.2
PyPDF2>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
